Fix file downloading of shorturl instead of page load. (#147)
The S3 static http bucket infers mimetype based on filename extension, and since the shorturl redirect generated a file with no extension, it caused shortcodes to serve binary data that triggers a download in the browser.  This does not show up with astro's `dev` or `preview` mode as the http server in both cases examines the file headers to determine mime type. 
Generate a html page using a `<meta>` tag do handle redirecting instead of using the Astro `redirect()` call.  This limits the solution to static site generation only, which is sufficient for now.