Structural changes to make `key:` fields mapping keys instead.
This makes the value of the `key:` field the actual key in an mapping.  This makes for easier consumption and lets the parse/editor guarentee keys are indeed unique.
In addition edited `abstract:` and `annotation:` fields to block style line folding.
Added quotes around bare `entrytype:` values since @ must be quoted.